+++ service/conversations/v1/service/configuration/api_op_fetch.go
@@ -14,6 +14,7 @@ type FetchConfigurationResponse struct {
 	DefaultChatServiceRoleSid         string `json:"default_chat_service_role_sid"`
 	DefaultConversationCreatorRoleSid string `json:"default_conversation_creator_role_sid"`
 	DefaultConversationRoleSid        string `json:"default_conversation_role_sid"`
+	ReachabilityEnabled               bool   `json:"reachability_enabled"`
 	URL                               string `json:"url"`
 }