+++ service/conversations/v1/service/configuration/api_op_update.go
@@ -13,6 +13,7 @@ type UpdateConfigurationInput struct {
 	DefaultChatServiceRoleSid         *string `form:"DefaultChatServiceRoleSid,omitempty"`
 	DefaultConversationCreatorRoleSid *string `form:"DefaultConversationCreatorRoleSid,omitempty"`
 	DefaultConversationRoleSid        *string `form:"DefaultConversationRoleSid,omitempty"`
+	ReachabilityEnabled               *bool   `form:"ReachabilityEnabled,omitempty"`
 }
 
 // UpdateConfigurationResponse defines the response fields for the updated service configuration
@@ -21,6 +22,7 @@ type UpdateConfigurationResponse struct {
 	DefaultChatServiceRoleSid         string `json:"default_chat_service_role_sid"`
 	DefaultConversationCreatorRoleSid string `json:"default_conversation_creator_role_sid"`
 	DefaultConversationRoleSid        string `json:"default_conversation_role_sid"`
+	ReachabilityEnabled               bool   `json:"reachability_enabled"`
 	URL                               string `json:"url"`
 }